header footer fix
