+++ src/resources/users/UserMusterList.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import Printable from '../../components/Printable'
-import { Box, Typography } from '@mui/material'
+import { Box, Table, TableFooter, TableHead, Typography } from '@mui/material'
 import {
   TextField,
   useListContext,
@@ -14,6 +14,8 @@ import SourceField from '../../components/SourceField'
 import * as constants from '../../constants'
 import ReportSignature from '../../components/ReportSignature'
 import { DateTime } from 'luxon'
+import { useConfigData } from '../../utils/useConfigData'
+import { Footer, Header } from '../../components/VaultLocationReport'
 
 interface Props {
   open: boolean
@@ -37,6 +39,7 @@ export default function UserMusterList<T extends User>(
   const { open, onClose } = props
   const { selectedIds } = useListContext()
   const userIds: number[] = selectedIds
+  const configData = useConfigData()
 
   return (
     <Printable open={open} onClose={onClose}>
@@ -44,78 +47,87 @@ export default function UserMusterList<T extends User>(
         {userIds.map((userId, index) => {
           return (
             <React.Fragment key={userId}>
-              <Box padding={'20px'} key={index}>
-                <Show<T>
-                  component={'div'}
-                  id={userId}
-                  resource={constants.R_USERS}
-                  actions={false}>
-                  <Title />
-                  <Typography
-                    fontSize='16px'
-                    variant='h5'
-                    textAlign='center'
-                    margin='10px'>
-                    {
-                      <Count
-                        resource={constants.R_ITEMS}
-                        sx={{ fontSize: '16px' }}
-                        filter={{ loanedTo: userId }}
-                      />
-                    }{' '}
-                    Items Currently Booked Out{' '}
-                    {DateTime.now().toFormat(constants.DATETIME_FORMAT)}
-                  </Typography>
-                </Show>
-                <ItemsReport
-                  filter={{ loanedTo: userId }}
-                  sx={{
-                    '.MuiTableCell-head span': {
-                      fontSize: '12px'
-                    }
-                  }}
-                  headStyle={{
-                    fontSize: '12px'
-                  }}>
-                  <TextField<Item>
-                    {...style}
-                    source='itemNumber'
-                    label='Item Number'
-                  />
-                  <SourceField<Item>
-                    textProps={{ ...style }}
-                    link='show'
-                    source='mediaType'
-                    reference={constants.R_MEDIA_TYPE}
-                    label='Media type'
-                  />
-                  <SourceField<Item>
-                    textProps={{ ...style }}
-                    source='protectiveMarking'
-                    reference={constants.R_PROTECTIVE_MARKING}
-                  />
-                  <DateField<Item>
-                    {...style}
-                    source='loanedDate'
-                    label='Loaned Date'
-                  />
-                  <TextField<Item>
-                    {...style}
-                    source='consecSheets'
-                    label='Consec/Sheets'
-                  />
-                </ItemsReport>
-                <ReportSignature>
-                  <Count
-                    resource={constants.R_ITEMS}
-                    {...style}
+              <Table>
+                <TableHead>
+                  <Header configData={configData} />
+                </TableHead>
+
+                <Box padding={'20px'} key={index}>
+                  <Show<T>
+                    component={'div'}
+                    id={userId}
+                    resource={constants.R_USERS}
+                    actions={false}>
+                    <Title />
+                    <Typography
+                      fontSize='16px'
+                      variant='h5'
+                      textAlign='center'
+                      margin='10px'>
+                      {
+                        <Count
+                          resource={constants.R_ITEMS}
+                          sx={{ fontSize: '16px' }}
+                          filter={{ loanedTo: userId }}
+                        />
+                      }{' '}
+                      Items Currently Booked Out{' '}
+                      {DateTime.now().toFormat(constants.DATETIME_FORMAT)}
+                    </Typography>
+                  </Show>
+                  <ItemsReport
                     filter={{ loanedTo: userId }}
-                  />
-                </ReportSignature>
-              </Box>
-              {selectedIds.length !== index + 1 && (
-                <div className='pagebreak' />
-              )}
+                    sx={{
+                      '.MuiTableCell-head span': {
+                        fontSize: '12px'
+                      }
+                    }}
+                    headStyle={{
+                      fontSize: '12px'
+                    }}>
+                    <TextField<Item>
+                      {...style}
+                      source='itemNumber'
+                      label='Item Number'
+                    />
+                    <SourceField<Item>
+                      textProps={{ ...style }}
+                      link='show'
+                      source='mediaType'
+                      reference={constants.R_MEDIA_TYPE}
+                      label='Media type'
+                    />
+                    <SourceField<Item>
+                      textProps={{ ...style }}
+                      source='protectiveMarking'
+                      reference={constants.R_PROTECTIVE_MARKING}
+                    />
+                    <DateField<Item>
+                      {...style}
+                      source='loanedDate'
+                      label='Loaned Date'
+                    />
+                    <TextField<Item>
+                      {...style}
+                      source='consecSheets'
+                      label='Consec/Sheets'
+                    />
+                  </ItemsReport>
+                  <ReportSignature>
+                    <Count
+                      resource={constants.R_ITEMS}
+                      {...style}
+                      filter={{ loanedTo: userId }}
+                    />
+                  </ReportSignature>
+                </Box>
+                {selectedIds.length !== index + 1 && (
+                  <div className='pagebreak' />
+                )}
+                <TableFooter>
+                  <Footer configData={configData} />
+                </TableFooter>
+              </Table>
             </React.Fragment>
           )
         })}
